requirements.txt
streamlit
openai
python-dotenv
fpdf
pandas

